Move view mixin to plugin.views
