+++ InvenTree/InvenTree/views.py
@@ -38,8 +38,6 @@ from part.models import PartCategory
 from common.models import InvenTreeSetting, ColorTheme
 from users.models import check_user_role, RuleSet
 
-from plugin.registry import registry
-
 from .forms import DeleteForm, EditUserForm, SetPasswordForm
 from .forms import SettingCategorySelectForm
 from .helpers import str2bool
@@ -56,37 +54,6 @@ def auth_request(request):
         return HttpResponse(status=200)
     else:
         return HttpResponse(status=403)
-
-
-class InvenTreePluginMixin:
-    """
-    Custom view mixin which adds context data to the view,
-    based on loaded plugins.
-
-    This allows rendered pages to be augmented by loaded plugins.
-
-    """
-
-    def get_plugin_panels(self, ctx):
-        """
-        Return a list of extra 'plugin panels' associated with this view
-        """
-
-        panels = []
-
-        for plug in registry.with_mixin('panel'):
-            panels += plug.render_panels(self, self.request, ctx)
-
-        return panels
-
-    def get_context_data(self, **kwargs):
-
-        ctx = super().get_context_data(**kwargs)
-
-        if settings.PLUGINS_ENABLED:
-            ctx['plugin_panels'] = self.get_plugin_panels(ctx)
-
-        return ctx
 
 
 class InvenTreeRoleMixin(PermissionRequiredMixin):
